requirements.txt
rembg[cpu]
fastapi
uvicorn
python-multipart
pillow
openai
u2net
rembg